Initial Version
I have finished the initial version including a README file.
Descriptions of the design and improvements can be found in the README
file. I need to add an installation guide to the README and address the
trailing adverb issue in the code. I also want to move the word lists
into external files (JSON or CSV would be acceptable formats).