Added determiners, fixed trailing adverbs
The poems generated now have determiners before every noun. These are
added after the word class sequence for a sentence is generated. The
list of states (word classes) is iterated through and for every noun a
backward pass is ran. During the backwards pass, the program searches
for the first state that is not an adjective and records it. Once this
state has been found, the determiner is placed after the recorded state.

Trailing adverbs have been fixed by checking if the next word is the
final word in the line, and if it is an adverb, not adding a full stop.